requirements.txt
Django==2.1.3
django-filter==2.0.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
Markdown==3.0.1
PyJWT==1.6.4
pytz==2018.7
